Changes to login and signout button, dashboard page, authenticaterouting
