+++ foi-web/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/dashboard.js
@@ -1,22 +1,8 @@
-import React, {useEffect} from 'react';
-import { useDispatch, useSelector } from "react-redux";
-import UserService from "../../../services/UserService";
-import { setUserAuth } from "../../../actions/bpmActions";
+import React from 'react';
 import DashboardImage from "../../../assets/FOI/images/DashboardImage.PNG"
 
-const Dashboard = React.memo((props) => { 
-  const dispatch = useDispatch();
-  useEffect(()=>{
-    console.log(JSON.stringify(props.store));
-    if(props.store){
-      UserService.initKeycloak(props.store, (err, res) => {
-        dispatch(setUserAuth(res.authenticated));
-      });
-    }
-  },[props.store, dispatch]);
-
-
-    return (      
+const Dashboard = React.memo(() => { 
+     return (      
         <div className="dashboard">
           <img src={DashboardImage} alt="Dashboard" />
         </div>
