+++ foi-web/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/index.js
@@ -1,3 +1,3 @@
-import dashboard from "./dashboard";
+import Dashboard from "./Dashboard";
 
-export default dashboard;
+export default Dashboard;
