+++ foi-web/forms-flow-ai/forms-flow-web/src/components/FOI/FOIBaseRouting.jsx
@@ -1,35 +1,34 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
-import {useSelector} from "react-redux";
-
-import FOIHeader from "./Header";
-import Home from "./Home";
-import Dashboard from "./Dashboard";
-import FOIFooter from "./Footer";
 import { ToastContainer } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css'
+import 'react-toastify/dist/ReactToastify.css';
+
+import Home from "./Home";
+import FOIHeader from "./Header";
+import FOIFooter from "./Footer";
+import FOIAuthenticateRouting from "./FOIAuthenticateRouting";
 
 
 const FOIBaseRouting = React.memo(({store}) => {
-    const isAuth = useSelector((state) => state.user.isAuthenticated);
-    console.log("Debug is authenticated "+ isAuth);
+  
     return (
       <>
-        <FOIHeader/>
-       
+              
               <ToastContainer />
               <Switch>
+
+                <Route path="/Dashboard">
+                    <FOIAuthenticateRouting store={store} />
+                </Route>
                
                 <Route exact path="/">
-                    <Home/>
+                    <FOIHeader /> 
+                    <Home />
+                    <FOIFooter />
                 </Route>
-                <Route exact path="/Dashboard">
-                    <Dashboard store={store}/>
-                </Route>
-              </Switch>
-              
+
+              </Switch>   
         
-        <FOIFooter/>
       </>
     );
   });
